# [IOTDB-632]Performance improve for Linear Fill