--- a/server/src/main/java/org/apache/iotdb/db/query/executor/fill/IFill.java
+++ b/server/src/main/java/org/apache/iotdb/db/query/executor/fill/IFill.java
@@ -21,12 +21,11 @@ package org.apache.iotdb.db.query.executor.fill;
 
 import org.apache.iotdb.db.exception.StorageEngineException;
 import org.apache.iotdb.db.exception.query.QueryProcessException;
-import org.apache.iotdb.db.exception.query.UnSupportedFillTypeException;
+import org.apache.iotdb.db.qp.physical.crud.RawDataQueryPlan;
 import org.apache.iotdb.db.query.context.QueryContext;
 import org.apache.iotdb.tsfile.file.metadata.enums.TSDataType;
 import org.apache.iotdb.tsfile.read.TimeValuePair;
 import org.apache.iotdb.tsfile.read.common.Path;
-import org.apache.iotdb.tsfile.read.filter.basic.Filter;
 
 import java.io.IOException;
 import java.util.Set;
@@ -47,14 +46,10 @@ public abstract class IFill {
   public abstract IFill copy();
 
   public abstract void configureFill(Path path, TSDataType dataType, long queryTime,
-      Set<String> sensors, QueryContext context)
-      throws StorageEngineException, QueryProcessException;
+      Set<String> sensors, QueryContext context);
 
-  public Filter getFilter() {
-    return constructFilter();
-  }
-
-  public abstract TimeValuePair getFillResult() throws IOException, UnSupportedFillTypeException;
+  public abstract TimeValuePair getFillResult()
+      throws IOException, QueryProcessException, StorageEngineException;
 
   public TSDataType getDataType() {
     return this.dataType;
@@ -68,5 +63,5 @@ public abstract class IFill {
     this.queryTime = queryTime;
   }
 
-  abstract Filter constructFilter();
+  abstract void constructFilter();
 }